Update README.md
In the 'What else can sdm do?' section, added a link to the Wiki page that explains appending lines to /etc/fstab.  I sometimes find it difficult to find the appropriate details for topics/concepts and have thought that it would be very useful to have such links right where the topic is introduced in the excellent overview README.  Rather than just ask for such a thing, I thought I would try to contribute in a small way.  I am pretty new to Github so I'm submitting this as a sort of test case.  It it's useful and properly done, I can add more such links as I work through my setup and use of sdm.